WINDUP-2457 Refactored 'Version' classes to avoid duplication
