+++ rules-java/tests/src/test/java/org/jboss/windup/rules/java/VersionTest.java
@@ -7,17 +7,17 @@ import org.junit.Test;
 public class VersionTest {
     @Test
     public void versionsCompareTest() {
-        Version version = Version.fromVersion("1.2.3").toVersion("3.2.1");
-        Assert.assertEquals(true, version.validate("2.2"));
-        Assert.assertEquals(true, version.validate("2.2.2.Final"));
-        Assert.assertEquals(true, version.validate("1.2.3"));
-        Assert.assertEquals(true, version.validate("3.2.1.Final"));
-        version = Version.fromVersion("1.0.0.RELEASE").toVersion("2.1.1.RELEASE");
-        Assert.assertEquals(true, version.validate("1.2.7.RELEASE"));
-        Assert.assertEquals(false, version.validate("2.2.7.RELEASE"));
-        Assert.assertEquals(true, version.validate("1.0.0.RELEASE"));
-        Assert.assertEquals(true, version.validate("1.0.0.Final"));
-        Assert.assertEquals(true, version.validate("2.1.1.RELEASE"));
-        Assert.assertEquals(true, version.validate("2.1.1"));
+        Version version = Version.fromVersion("1.2.3").to("3.2.1");
+        Assert.assertTrue(version.validate("2.2"));
+        Assert.assertTrue(version.validate("2.2.2.Final"));
+        Assert.assertTrue(version.validate("1.2.3"));
+        Assert.assertTrue(version.validate("3.2.1.Final"));
+        version = Version.fromVersion("1.0.0.RELEASE").to("2.1.1.RELEASE");
+        Assert.assertTrue(version.validate("1.2.7.RELEASE"));
+        Assert.assertFalse(version.validate("2.2.7.RELEASE"));
+        Assert.assertTrue(version.validate("1.0.0.RELEASE"));
+        Assert.assertTrue(version.validate("1.0.0.Final"));
+        Assert.assertTrue(version.validate("2.1.1.RELEASE"));
+        Assert.assertTrue(version.validate("2.1.1"));
     }
 }
